Removed old builder module
